+++ src/draw/ft_distance.c
@@ -6,13 +6,13 @@
 /*   By: ryoshio- <ryoshio-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/10 20:52:30 by ryoshio-          #+#    #+#             */
-/*   Updated: 2023/03/10 21:00:32 by ryoshio-         ###   ########.fr       */
+/*   Updated: 2023/03/29 00:45:41 by ryoshio-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/cub3d.h"
 
-float	ft_distance(float x1, float y1, float x2, float y2)
+double	ft_distance(double x1, double y1, double x2, double y2)
 {
 	return (sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2)));
 }
