+++ src/draw/ft_size_wall.c
@@ -6,7 +6,7 @@
 /*   By: ryoshio- <ryoshio-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 02:53:28 by ryoshio-          #+#    #+#             */
-/*   Updated: 2023/03/28 23:48:42 by ryoshio-         ###   ########.fr       */
+/*   Updated: 2023/03/29 00:45:41 by ryoshio-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 int ft_size_wall(t_cub *cub)
 {
-	float tmp;
+	double tmp;
 	if(cub->play.dist == 0)
 		return (0);
 	tmp = TEXTURE_SIZE * WIN_SIZE_Y / cub->play.dist;
@@ -28,7 +28,7 @@ int ft_size_wall(t_cub *cub)
 
 /*
 
-	float tmp;
+	double tmp;
 	if(cub->play.dist == 0)
 		return (0);
 	tmp = TEXTURE_SIZE / cub->play.dist * floor( WIN_SIZE_Y/(2* tan(PI/6)));
